// frontend/src/model/pycon/payments.ts
import { CreateRegistration } from './registrations';

export interface GetTransactionDetailsOut {
  ticket_price: number;
  transaction_fee: number;
  total_price: number;
  platform_fee: number;
}

export interface TransactionDetails {
  ticket_price: number;
  payment_method: PaymentMethod;
  payment_channel: eWalletChannelCode | DirectDebitChannelCode;
  platform_fee: number | null;
}

export type PaymentMethod = 'DIRECT_DEBIT' | 'E_WALLET';

export type eWalletChannelCode = 'GCASH' | 'PAYMAYA';

export type DirectDebitChannelCode = 'BPI_DIRECT_DEBIT' | 'UBP_DIRECT_DEBIT' | 'RCBC_DIRECT_DEBIT'
// | 'CHINABANK';

export type PaymentChannel = eWalletChannelCode | DirectDebitChannelCode;

interface PaymentIn {
  amount: number;
  successReturnUrl: string;
  failureReturnUrl: string;
  cancelReturnUrl?: string;
  eventId: string;
  registrationData: Omit<CreateRegistration, 'amountPaid' | 'transactionId'>;
}

export interface EWalletPaymentIn extends PaymentIn {
  referenceId: string;
  channelCode: eWalletChannelCode;
}

export interface DirectDebitPaymentIn extends PaymentIn {
  email: string;
  givenNames: string;
  surname: string;
  channelCode: DirectDebitChannelCode;
}

export interface PaymentRequestOut {
  createDate: string;
  paymentUrl: string;
  paymentRequestId: string;
  referenceId: string;
}

export interface PaymentRequestOutV2 {
  payment_request_id: string,
  reference_id: string,
  status: string,
  actions: Array<{
    type: string;
    value: string;
    descriptor: string;
  }>;
  created: string;
}